fix(ci): use env var for database dsn in tests
tests had hardcoded postgresql://localhost/synapto without credentials,
which fails in ci where the postgres service requires a password.
extracted shared pg/redis fixtures into tests/conftest.py that reads
SYNAPTO_PG_DSN and SYNAPTO_REDIS_URL from environment with local
fallback defaults.
