+++ tests/unit/test_db.py
@@ -2,31 +2,9 @@
 
 from __future__ import annotations
 
-import pytest
-
 from synapto.db.migrations import get_schema_version, run_migrations
 from synapto.db.postgres import PostgresClient
 from synapto.db.redis_cache import RedisCache
-
-DSN = "postgresql://localhost/synapto"
-REDIS_URL = "redis://localhost:6379/1"  # use db 1 for tests
-
-
-@pytest.fixture
-async def pg():
-    client = PostgresClient(DSN, min_size=1, max_size=2)
-    await client.connect()
-    yield client
-    await client.close()
-
-
-@pytest.fixture
-async def cache():
-    c = RedisCache(url=REDIS_URL, prefix="synapto_test")
-    await c.connect()
-    yield c
-    await c.flush_prefix()
-    await c.close()
 
 
 class TestPostgresClient:
